Started coding.
Reading and writing to registers implemented.
The circuit is now on rev 1.1, but there were some mistakes on this one to, hehe.
The dip-switch is NOT the same size as a 16p soic, a little wider (bodgedd by bending pins under on one side (the one thats just all vcc anyways, as it can be bridged if you're struggeling with the jank solder)).
The socket also has some mechanical solder points on the ends, that wasnt on my footprint. They are also thinner and longer than thought. The mechanical solder on one side could short some pins on the 165, but i simply put a piece of kapton tape acrross the pins. Circuit works otherwise. Perhaps a little hot. Maybe not ideal to drive LEDs directly from 595's, but they can probably handle it